Move stateless ops out of contrib
