+++ tensorflow/python/kernel_tests/random/BUILD
@@ -91,6 +91,21 @@ cuda_py_test(
 )
 
 cuda_py_test(
+    name = "stateless_random_ops_test",
+    size = "medium",
+    srcs = ["stateless_random_ops_test.py"],
+    additional_deps = [
+        "//third_party/py/numpy",
+        "//tensorflow/python:array_ops",
+        "//tensorflow/python:client_testlib",
+        "//tensorflow/python:framework",
+        "//tensorflow/python:platform",
+        "//tensorflow/python:random_ops",
+        "//tensorflow/python:stateless_random_ops",
+    ],
+)
+
+cuda_py_test(
     name = "random_gamma_test",
     size = "medium",
     srcs = ["random_gamma_test.py"],
